+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [0.8.4] - 2018-04-18
 
 ### Fixed
 
@@ -255,7 +255,8 @@ Initial release
 
 
 
-[Unreleased]: https://github.com/sdispater/poetry/compare/0.8.3...master
+[Unreleased]: https://github.com/sdispater/poetry/compare/0.8.4...master
+[0.8.4]: https://github.com/sdispater/poetry/releases/tag/0.8.4
 [0.8.3]: https://github.com/sdispater/poetry/releases/tag/0.8.3
 [0.8.2]: https://github.com/sdispater/poetry/releases/tag/0.8.2
 [0.8.1]: https://github.com/sdispater/poetry/releases/tag/0.8.1